# server/openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "API for adding, retrieving and removing centers and events",
    "version": "1",
    "title": "EventsManager API",
    "contact": {
      "email": "adegold71@gmail.com"
    }
  },
  "basePath": "/api/v1/",
...
  "schemes": [
    "https", "http"
  ],
  "paths": {
...
     "/centers": {
      "post": {
        "tags": [
          "Centers"
        ],
        "summary": "Creates a new event center",
        "operationId": "CenterController.addCenter",
        "consumes" : [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "adminSecurity": []
          }
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Details used to create new event center",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SubmitCenter"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful request",
            "schema": {
              "$ref": "#/definitions/Center"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Authentication error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "Admin not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Centers"
        ],
        "summary": "Fetches all existing event centers",
        "operationId": "CenterController.fetchCenters",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name": "filter",
            "in": "query",
            "description": "fetched centers will have this in their names or addresses",
            "required": false,
            "type": "string"
          },
          {
            "name": "facility",
            "in": "query",
            "description": "fetched centers will have this in their list of facilities",
            "required": false,
            "type": "string"
          },
          {
            "name": "capacity",
            "in": "query",
            "description": "fetched centers can contain this or more gursts",
            "required": false,
            "type": "integer"
          },
          {
            "name": "offset",
            "in": "query",
            "description": "offset the fetched centers by this value",
            "required": false,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "only centers within this limit will be fetched",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "schema": {
              "items": {
                "$ref": "#/definitions/Center"
              }
            }
          }
        }
      }
    },
    "/centers/{id}": {
      "put": {
        "tags": [
          "Centers"
        ],
        "summary": "Modifies an existing event center",
        "operationId": "CenterController.modifyCenter",
        "consumes" : [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "adminSecurity": []
          }
        ],
        "parameters": [
          { 
            "name": "id",
            "in": "path",
            "description": "id of the center to be modified",
            "required": true,
            "type": "integer"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Details used to edit existing event center",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SubmitCenter"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful request",
            "schema": {
              "$ref": "#/definitions/Center"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "401": {
            "description": "Authentication error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "404": {
            "description": "Admin not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Centers"
        ],
        "summary": "Fetches an existing event center",
        "operationId": "CenterController.fetchCenter",
        "produces": [
          "application/json"
        ],
        "parameters": [
          { 
            "name": "id",
            "in": "path",
            "description": "id of the center to be fetched",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful request",
            "schema": {
              "$ref": "#/definitions/Center"
            }
          },
          "404": {
            "description": "Center not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "userSecurity": {
      "type": "apiKey",
      "name": "token",
      "in": "query"
    },
    "adminSecurity": {
      "type": "apiKey",
      "name": "token",
      "in": "query"
    }
  },
  "definitions": {
...
    "SubmitCenter": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "facilities": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "cost": {
          "type": "integer"
        },
        "capacity": {
          "type": "integer"
        },
        "images": {
          "type": "string"
        }
      },
      "required": ["name", "description", "address", "cost", "capacity", "facilities", "images"]
    },
    "Center": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "center": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "facilities": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "address": {
              "type": "string"
            },
            "id": {
              "type": "integer"
            },
            "cost": {
              "type": "integer"
            },
            "capacity": {
              "type": "integer"
            },
            "images": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "required": ["name", "description", "address", "cost", "capacity", "facilities", "id", "images"]
        }
      },
      "required": ["status", "center"]
    },
...
    "Error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      },
      "required": [
        "status", "error"
      ]
    },
...
  }
}